// src/generated/prisma/models/WorkExperience.ts
export type WorkExperienceModel = runtime.Types.Result.DefaultSelection<Prisma.$WorkExperiencePayload>
// ...
export type $WorkExperiencePayload<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  name: "WorkExperience"
  objects: {
    profile: Prisma.$ProfilePayload<ExtArgs>
  }
  scalars: runtime.Types.Extensions.GetPayloadResult<{
    id: string
    profileId: string
    company: string
    position: string
    startDate: Date
    endDate: Date | null
    description: string
    technologies: string[]
  }, ExtArgs["result"]["workExperience"]>
  composites: {}
}