// src/generated/prisma/models/Education.ts
export type EducationModel = runtime.Types.Result.DefaultSelection<Prisma.$EducationPayload>
// ...
export type $EducationPayload<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  name: "Education"
  objects: {
    profile: Prisma.$ProfilePayload<ExtArgs>
  }
  scalars: runtime.Types.Extensions.GetPayloadResult<{
    id: string
    profileId: string
    institution: string
    degree: string
    startDate: Date
    endDate: Date | null
  }, ExtArgs["result"]["education"]>
  composites: {}
}