// src/generated/prisma/models/Skill.ts
export type SkillModel = runtime.Types.Result.DefaultSelection<Prisma.$SkillPayload>
// ...
export type $SkillPayload<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  name: "Skill"
  objects: {
    profile: Prisma.$ProfilePayload<ExtArgs>
  }
  scalars: runtime.Types.Extensions.GetPayloadResult<{
    id: string
    profileId: string
    name: string
    category: string | null
    level: number | null
  }, ExtArgs["result"]["skill"]>
  composites: {}
}