// src/generated/prisma/models/Certificate.ts
export type CertificateModel = runtime.Types.Result.DefaultSelection<Prisma.$CertificatePayload>
// ...
export type $CertificatePayload<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  name: "Certificate"
  objects: {
    profile: Prisma.$ProfilePayload<ExtArgs>
  }
  scalars: runtime.Types.Extensions.GetPayloadResult<{
    id: string
    profileId: string
    name: string
    issuer: string
    issueDate: Date | null
    expiryDate: Date | null
    description: string | null
    imageUrl: string | null
    createdAt: Date
  }, ExtArgs["result"]["certificate"]>
  composites: {}
}